+++ packages/terra-content-container/src/ContentContainer.jsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames/bind';
-import 'terra-base/lib/baseStyles';
 import Scroll from 'terra-scroll';
 import styles from './ContentContainer.module.scss';