+++ packages/terra-demographics-banner/src/DemographicsBanner.jsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import 'terra-base/lib/baseStyles';
 import DemographicsBannerDisplay from './DemographicsBannerDisplay';
 
 const propTypes = {