+++ packages/terra-form-radio/src/RadioField.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames/bind';
 import uniqueid from 'lodash.uniqueid';
-import 'terra-base/lib/baseStyles';
+
 import styles from './RadioField.module.scss';
 
 const cx = classNames.bind(styles);