+++ packages/terra-table/src/SingleSelectableRows.jsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import 'terra-base/lib/baseStyles';
 import SelectableTableRows from './SelectableTableRows';
 
 const propTypes = {