+++ packages/terra-form-checkbox/CHANGELOG.md
@@ -6,6 +6,9 @@ Unreleased
 ### Added
 * Added aria-describedby to better announce legend, help text, and error text in the Radio Field component
 
+### Removed
+* Removed peer dependency on terra-base
+* Removed baseStyles import from terra-base
 
 3.6.0 - (March 21, 2019)
 ------------------